backend/requirements.txt
Django==3.2.7
asgiref==3.4.1
pytz==2021.1
sqlparse==0.4.2
django-ckeditor
djangorestframework==3.12.4
djoser==2.1.0
django-filter==2.4.0
django-extra-fields==3.0.2
pytest==6.1.2
gunicorn==20.1.0
psycopg2==2.8.6